+++ src/test/java/com/kt/controller/admin/AdminControllerTest.java
@@ -31,9 +31,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @ActiveProfiles("test")
 public class AdminControllerTest extends AbstractRestDocsTest {
 
-    private static final String ADMIN_URL_PREFIX = "/admins";
-    private static final String ADMIN_LOGIN_ID  = "PasswordTest123";
-    private static final String ADMIN_PASSWORD  = "PasswordTest123!";
+    private static final String ADMIN_URL_PREFIX = "/super-admin";
+    private static final String ADMIN_LOGIN_ID  = "superadmin123";
+    private static final String ADMIN_PASSWORD  = "SuperAdmin123!";
 
     @MockitoBean
     private StringRedisTemplate stringRedisTemplate;
@@ -100,7 +100,7 @@ public class AdminControllerTest extends AbstractRestDocsTest {
                                     request,
                                     HttpMethod.POST,
                                     objectMapper
-                            ).with(jwtAdmin())
+                            ).with(jwtSuperAdmin())
                     )
                     .andExpect(status().isCreated())
                     .andDo(
@@ -126,7 +126,7 @@ public class AdminControllerTest extends AbstractRestDocsTest {
                             null,
                             HttpMethod.GET,
                             objectMapper
-                    ).with(jwtAdmin())
+                    ).with(jwtSuperAdmin())
             )
                     .andExpect(status().isOk())
                     .andExpect(jsonPath("$.data").isArray())
@@ -160,7 +160,7 @@ public class AdminControllerTest extends AbstractRestDocsTest {
                                     HttpMethod.GET,
                                     objectMapper,
                                     adminId
-                            ).with(jwtAdmin())
+                            ).with(jwtSuperAdmin())
                     )
                     .andExpect(status().isOk())
                     .andExpect(jsonPath("$.data.id").value(adminId))
@@ -201,7 +201,7 @@ public class AdminControllerTest extends AbstractRestDocsTest {
                                     HttpMethod.PATCH,
                                     objectMapper,
                                     adminId
-                            ).with(jwtAdmin())
+                            ).with(jwtSuperAdmin())
                     )
                     .andExpect(status().isOk())
                     .andExpect(jsonPath("$.data.id").value(adminId))
@@ -234,7 +234,7 @@ public class AdminControllerTest extends AbstractRestDocsTest {
                                     HttpMethod.DELETE,
                                     objectMapper,
                                     adminId
-                            ).with(jwtAdmin())
+                            ).with(jwtSuperAdmin())
                     )
                     .andExpect(status().isNoContent())
                     .andDo(
@@ -261,7 +261,7 @@ public class AdminControllerTest extends AbstractRestDocsTest {
                             HttpMethod.PATCH,
                             objectMapper,
                             adminId
-                    ).with(jwtAdmin())
+                    ).with(jwtSuperAdmin())
             )
                     .andExpect(status().isNoContent())
                     .andDo(
